pyproject.toml
[project]
name = "fl-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.36.20",
    "dj-database-url>=2.3.0",
    "django>=5.1.6",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-s3-storage>=0.15.0",
    "djangorestframework>=3.15.2",
    "feedparser>=6.0.11",
    "gunicorn>=23.0.0",
    "pillow>=11.1.0",
    "pre-commit>=4.1.0",
    "psycopg>=3.2.4",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "whitenoise>=6.9.0",
    "pytest>=8.3.0",
    "pytest-django>=4.9.0",
    "pytest-cov>=5.0.0",
]

[tool.ruff]
target-version = "py310"
line-length = 100
extend-exclude = ["migrations", "__pycache__"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "B",  # bugbear
    "I",  # import sorting
    "UP", # pyupgrade
    "DJ", # django rules
]
ignore = [
    "E501",  # line length (handled by formatter and we set 100)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
docstring-code-format = true
line-ending = "lf"
